Create demo UI for updateObject function
This commit adds some demo UI (React components) that allows the user to
test the updateObject function and its functionality. To store the data,
this commit also adds firebase dependency and uses it. Additionally,
this commit implements ObjectService, which handles CRUD operations with
objects (to database). This commit also fixes some typescript issues in
the project.