Add JsonInput, JsonOutput, and Improve ConfirmDialog
With this commit, I added 2 new components:
- JsonInput -> where the user can input JSON and it is auto validated
- JsonOutput -> displays formatted JSON on the UI

To do this, I added new dependencies to `package.json`.

I also refactored and improved the code of `ConfirmDialog` so there is a
new prop `type` that is responsible for selecting the corresponding
component. Right now, there are 3 options for ConfirmDialog:
1. Without an input field
2. Input JSON field
3. Text input field

This is done to improve UI/UX of this small app.